chore: tag, menu api, 관련 query 훅 관심사 분리
