+++ src/query/shopManage/menuQueries.js
@@ -1,31 +1,5 @@
 import { useQuery, useMutation } from '@tanstack/react-query';
-import { shopManageService } from '../api/services/shopManageService';
-
-/**
- * 태그(카테고리) 목록 조회 훅
- */
-export const useShopManageTags = () => {
-  return useQuery({
-    queryKey: ['shop-manage-tags'],
-    queryFn: () => shopManageService.getTags(),
-  });
-};
-
-/**
- * 태그(카테고리) 생성 훅
- */
-export const useCreateShopTag = () => {
-  return useMutation({
-    mutationFn: (name) => shopManageService.createTag(name),
-    onSuccess: () => {
-      alert('태그가 생성되었습니다.');
-    },
-    onError: (error) => {
-      console.error('태그 생성 실패:', error);
-      alert('태그 생성 중 오류가 발생했습니다.');
-    },
-  });
-};
+import { menuService } from '../../api/services/menuService';
 
 /**
  * 메뉴 목록 조회 훅
@@ -33,7 +7,7 @@ export const useCreateShopTag = () => {
 export const useShopManageMenus = (shopId) => {
   return useQuery({
     queryKey: ['shop-manage-menus', shopId],
-    queryFn: () => shopManageService.getMenus(shopId),
+    queryFn: () => menuService.getMenus(shopId),
     enabled: !!shopId,
   });
 };
@@ -44,7 +18,7 @@ export const useShopManageMenus = (shopId) => {
 export const useCreateShopMenu = () => {
   return useMutation({
     mutationFn: ({ shopId, name, description, sortOrder }) =>
-      shopManageService.createMenu(shopId, { name, description, sortOrder }),
+      menuService.createMenu(shopId, { name, description, sortOrder }),
     onSuccess: () => {
       alert('메뉴가 생성되었습니다.');
     },
@@ -66,7 +40,7 @@ export const useCreateShopMenu = () => {
 export const useUpdateShopMenu = () => {
   return useMutation({
     mutationFn: ({ shopId, menuId, name, description, sortOrder }) =>
-      shopManageService.updateMenu(shopId, menuId, { name, description, sortOrder }),
+      menuService.updateMenu(shopId, menuId, { name, description, sortOrder }),
     onSuccess: () => {
       alert('메뉴가 수정되었습니다.');
     },
@@ -82,7 +56,7 @@ export const useUpdateShopMenu = () => {
  */
 export const useDeactivateShopMenu = () => {
   return useMutation({
-    mutationFn: ({ shopId, menuId }) => shopManageService.deactivateMenu(shopId, menuId),
+    mutationFn: ({ shopId, menuId }) => menuService.deactivateMenu(shopId, menuId),
     onSuccess: () => {
       alert('메뉴가 비활성화되었습니다.');
     },
@@ -98,7 +72,7 @@ export const useDeactivateShopMenu = () => {
  */
 export const useLinkMenuTag = () => {
   return useMutation({
-    mutationFn: ({ shopId, menuId, tagId }) => shopManageService.linkMenuTag(shopId, menuId, tagId),
+    mutationFn: ({ shopId, menuId, tagId }) => menuService.linkMenuTag(shopId, menuId, tagId),
     onSuccess: () => {
       alert('태그가 연결되었습니다.');
     },
